setup TCR for deploy + rearranged
