+++ backend/src/server.ts
@@ -9,9 +9,9 @@ import scoreRoutes from "@/routes/score.route"
 dotenv.config();
 const app = express();
 const PORT = process.env.PORT || 5000;
-
+  
 app.use(cors({
-  origin: process.env.FRONTEND_URL,
+  origin: process.env.FRONTEND_URL || "http://localhost:5173",
   credentials: true
 }));
 app.use(express.json()); // Parse JSON
