+++ ts-code-runner/src/config.ts
@@ -1,0 +1,5 @@
+// ts-code-runner\src\config.ts
+import path from "path";
+
+export const TEMP_DIR = path.join(__dirname, "temp");
+export const RUNTIME_TIMEOUT_MS = 3000;
