+++ ts-code-runner/src/services/compile.service.ts
@@ -1,4 +1,4 @@
-// ts-code-runner\services\compile.service.ts
+// ts-code-runner\src\services\compile.service.ts
 import { exec } from "child_process";
 
 export function compileCode(cppFile: string, exeFile: string): Promise<void> {
